HW_SVOY_6_4

+++ HW/Test3_7.java
@@ -2,7 +2,7 @@ public class Test3_7 {
     public static void main(String [] args) {
         for (char i = 0; i <= 47; i++) {
             if (((int) i) % 2 != 0) {
-                System.out.printf("%6s%6s", ((int) i), i);
+                System.out.printf("%6s%6s", i, i);
                 System.out.println();
             }
         }
